+++ examples/resources/streamkap_source_oracleaws/basic.tf
@@ -16,7 +16,7 @@ resource "streamkap_source_oracleaws" "example" {
   table_include_list  = "HR.EMPLOYEES,HR.DEPARTMENTS"
 
   # Signal and heartbeat tables (required)
-  signal_data_collection_schema_or_database    = "STREAMKAP"
+  signal_data_collection_schema_or_database    = "STREAMKAP.STREAMKAP_SIGNAL"
   heartbeat_data_collection_schema_or_database = "STREAMKAP"
 }